[BE] refactor(#268): 피드백반영
