+++ backend/src/test/java/com/example/backend/study/api/service/member/StudyMemberServiceTest.java
@@ -9,19 +9,19 @@ import com.example.backend.common.exception.member.MemberException;
 import com.example.backend.domain.define.account.user.User;
 import com.example.backend.domain.define.account.user.repository.UserRepository;
 import com.example.backend.domain.define.event.FcmFixture;
-import com.example.backend.domain.define.fcmToken.FcmToken;
-import com.example.backend.domain.define.fcmToken.repository.FcmTokenRepository;
+import com.example.backend.domain.define.fcm.FcmToken;
+import com.example.backend.domain.define.fcm.repository.FcmTokenRepository;
 import com.example.backend.domain.define.study.info.StudyInfo;
 import com.example.backend.domain.define.study.info.StudyInfoFixture;
-import com.example.backend.domain.define.study.info.listener.StudyEventListener;
-import com.example.backend.domain.define.study.info.listener.event.ApplyApproveRefuseMemberEvent;
-import com.example.backend.domain.define.study.info.listener.event.ApplyMemberEvent;
+import com.example.backend.domain.define.fcm.listener.StudyEventListener;
+import com.example.backend.domain.define.study.info.event.ApplyApproveRefuseMemberEvent;
+import com.example.backend.domain.define.study.info.event.ApplyMemberEvent;
 import com.example.backend.domain.define.study.info.repository.StudyInfoRepository;
 import com.example.backend.domain.define.study.member.StudyMember;
 import com.example.backend.domain.define.study.member.StudyMemberFixture;
 import com.example.backend.domain.define.study.member.constant.StudyMemberStatus;
-import com.example.backend.domain.define.study.member.listener.MemberEventListener;
-import com.example.backend.domain.define.study.member.listener.event.ResignMemberEvent;
+import com.example.backend.domain.define.fcm.listener.MemberEventListener;
+import com.example.backend.domain.define.study.member.event.ResignMemberEvent;
 import com.example.backend.domain.define.study.member.repository.StudyMemberRepository;
 import com.example.backend.domain.define.study.todo.StudyTodoFixture;
 import com.example.backend.domain.define.study.todo.info.StudyTodo;
@@ -31,7 +31,6 @@ import com.example.backend.domain.define.study.todo.repository.StudyTodoReposito
 import com.example.backend.study.api.controller.member.request.ApplyMemberMessageRequest;
 import com.example.backend.study.api.controller.member.response.StudyMemberApplyListAndCursorIdxResponse;
 import com.example.backend.study.api.controller.member.response.StudyMembersResponse;
-import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
